+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 - [x] Updated format of .sprite JSON to use name-file pairs
 - [x] Generated file names all contain *.sprite* in them
+- [x] Support drag 'n drop of image files onto .sprite file
 - [ ] Command to update all sprites in solution
 - [ ] Call Web Compiler to recompile when sprite updated
-- [ ] Support drag 'n drop of image files onto .sprite file
 
 Features that have a checkmark are complete and available for
 download in the